backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
werkzeug==2.3.7
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
pillow==10.0.0
opencv-python-headless==4.8.0.76
ultralytics==8.0.196
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
pyyaml==6.0.1